update (knora-ui): renamed project from knora-gui into knora-ui; docs (README): update README files in core service
