+++ angular.json
@@ -3,7 +3,7 @@
   "version": 1,
   "newProjectRoot": "projects",
   "projects": {
-    "knora-gui": {
+    "knora-ui": {
       "root": "",
       "sourceRoot": "src",
       "projectType": "application",
@@ -13,7 +13,7 @@
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
-            "outputPath": "dist/knora-gui",
+            "outputPath": "dist/knora-ui",
             "index": "src/index.html",
             "main": "src/main.ts",
             "polyfills": "src/polyfills.ts",
@@ -52,18 +52,18 @@
         "serve": {
           "builder": "@angular-devkit/build-angular:dev-server",
           "options": {
-            "browserTarget": "knora-gui:build"
+            "browserTarget": "knora-ui:build"
           },
           "configurations": {
             "production": {
-              "browserTarget": "knora-gui:build:production"
+              "browserTarget": "knora-ui:build:production"
             }
           }
         },
         "extract-i18n": {
           "builder": "@angular-devkit/build-angular:extract-i18n",
           "options": {
-            "browserTarget": "knora-gui:build"
+            "browserTarget": "knora-ui:build"
           }
         },
         "test": {
@@ -97,7 +97,7 @@
         }
       }
     },
-    "knora-gui-e2e": {
+    "knora-ui-e2e": {
       "root": "e2e/",
       "projectType": "application",
       "architect": {
@@ -105,7 +105,7 @@
           "builder": "@angular-devkit/build-angular:protractor",
           "options": {
             "protractorConfig": "e2e/protractor.conf.js",
-            "devServerTarget": "knora-gui:serve"
+            "devServerTarget": "knora-ui:serve"
           }
         },
         "lint": {
@@ -200,5 +200,5 @@
       }
     }
   },
-  "defaultProject": "knora-gui"
+  "defaultProject": "knora-ui"
 }
