+++ deprecated-angular-cli.json
@@ -1,7 +1,7 @@
 {
   "$schema": "./node_modules/@angular/cli/lib/config/schema.json",
   "project": {
-    "name": "knora-gui"
+    "name": "knora-ui"
   },
   "apps": [
     {
