+++ docs/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html><html lang="en"><head>
   <meta charset="utf-8">
   <title>Demo and test app for Knora gui modules</title>
-  <base href="/Knora-gui/">
+  <base href="/Knora-ui/">
 
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="icon" type="image/x-icon" href="favicon.ico">
