+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "knora-gui",
+  "name": "knora-ui",
   "version": "0.0.0",
   "license": "none",
   "scripts": {
