+++ projects/knora/core/package.json
@@ -4,7 +4,7 @@
   "description": "Knora gui core module",
   "repository": {
     "type": "git",
-    "url": "https://github.com/dhlab-basel/Knora-gui.git"
+    "url": "https://github.com/dhlab-basel/Knora-ui.git"
   },
   "keywords": [
     "angular",
