+++ projects/knora/core/src/lib/declarations/core.config.ts
@@ -1,7 +1,7 @@
 import {JsonObject, JsonProperty} from 'json2typescript';
 
 /**
- * Knora-gui core configuration with the server definitions of:
+ * Knora-ui core configuration with the server definitions of:
  *  - api: URL of data service e.g. knora: http://localhost:3333
  *  - media: URL of media server service e.g. sipi: http://localhost:1024
  *  - gui: URL of the app e.g. salsah: http://localhost:4200
