+++ projects/knora/core/src/lib/services/projects/projects.service.spec.ts
