+++ src/app/knora-ui-examples/core-demo/core-demo.component.spec.ts
