+++ src/app/knora-ui-examples/core-demo/projects/projects.component.ts
