+++ src/app/knora-ui-examples/progress-indicator-demo/progress-indicator-demo.component.html
