+++ src/app/landing-page/demo-intro/demo-intro.component.html
@@ -1,1 +1,1 @@
-<h3>This is the area where you can test the Knora-gui modules. Select one from the left hand side</h3>
+<h3>This is the area where you can test the Knora-ui modules. Select one from the left hand side</h3>
