+++ src/app/landing-page/partials/main-header/main-header.component.html
@@ -21,7 +21,7 @@
 
   <span class="fill-remaining-space"></span>
 
-  <a href="https://github.com/dhlab-basel/Knora-gui" mat-button target="_blank">
+  <a href="https://github.com/dhlab-basel/Knora-ui" mat-button target="_blank">
       <mat-icon svgIcon="github-icon"></mat-icon>
       GitHub
   </a>
